// src/lib/moetruyen/model/postV2ChaptersByIdPageAccess200DataPagesItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Moetruyen Public API
 * Unofficial read-only REST API for [MoeTruyen](https://moetruyen.net/).

## Quick start

Use the `/v2` routes for the current stable surface.

## Request notes

- All API endpoints have a global rate limit of 7 requests per second per IP.
- Include a valid `Origin` header such as `https://suicaodex.com` or `https://moetruyen.net` when making browser-like requests.

## Versioning

- From 01/04/2026, v1 routes are deprecated, pleade use v2 instead.

## Repositories

- API repo: [TNTKien/moetruyen-public-api](https://github.com/TNTKien/moetruyen-public-api)
- Original site repo: [dex593/web1 (MoeTruyen)](https://github.com/dex593/web1)
 * OpenAPI spec version: 0.2.0
 */
import type { PostV2ChaptersByIdPageAccess200DataPagesItemGrant } from "./postV2ChaptersByIdPageAccess200DataPagesItemGrant";

export type PostV2ChaptersByIdPageAccess200DataPagesItem = {
  /**
   * @minimum 0
   * @maximum 9007199254740991
   */
  pageIndex: number;
  /**
   * @maximum 9007199254740991
   * @exclusiveMinimum 0
   */
  pageNumber: number;
  /** @minLength 1 */
  storageKey: string;
  downloadUrl: string;
  grant: PostV2ChaptersByIdPageAccess200DataPagesItemGrant;
};